+++ crates/julian-cli/CHANGELOG.md
@@ -2,6 +2,7 @@ v0.7.0 (in development)
 -----------------------
 - Add URL of GitHub repository to `--help` output
 - Increased MSRV to 1.88
+- Don't panic when writing to stdout fails due to a broken pipe
 
 v0.6.2 (2024-12-19)
 -------------------